% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{chat}
\alias{chat}
\title{Runs databot}
\usage{
chat(new_session = FALSE)
}
\arguments{
\item{new_session}{Logical. If \code{TRUE}, starts a new chat session. If \code{FALSE}
(the default), restores the previous chat session (only within the current
R process).}
}
\description{
Runs databot
}
