% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{confusion_matrix}
\usage{
confusion_matrix(test_y, predictions, margin = 1)
}
\arguments{
\item{test_y}{column from test dataframe you are trying to predict}

\item{predictions}{vector containing predictions}

\item{margin}{(default : 1) parameter in the prop.table() function from base R}
}
\value{
Returns a table in the form of a confusion matrix
}
\description{
Simplifies making a confusion matrix
}
\examples{
Predicting default in the Default dataset from the ISLR2 package

regression <- glm(default ~ income + balance, data = train, family = 'binomial')
predictions <- predict(regression, newdata = test, type = 'response') > 0.5)
confusion_matrix(test$default, predictions)

}
