% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databot-package.R
\docType{package}
\name{databot-package}
\alias{databot}
\alias{databot-package}
\title{databot: A Shiny App for Data Analysis with AI}
\description{
A Shiny application that provides an AI-powered interface for data analysis.
}
\author{
\strong{Maintainer}: Joe Cheng \email{joe@posit.co}

}
\keyword{internal}
