% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{knn}
\alias{knn}
\title{knn function}
\usage{
knn(x0, x, y, k)
}
\arguments{
\item{x0}{observation you are trying to predict}

\item{x}{observations used to predict}

\item{y}{prediction values corresponding to observations in x}

\item{k}{the k nearest neighbours to consider}
}
\value{
the predicted value corresponding to x0
}
\description{
knn function
}
\examples{
predicting horsepower based on miles per gallon in the ISLR2::Auto dataset

prediction <- knn(20, Auto$mpg, Auto$horsepower, 3)

}
