% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_test.R
\name{knn_test}
\alias{knn_test}
\title{knn_test}
\usage{
knn_test(test_x0, train_x, train_y, k)
}
\arguments{
\item{test_x0}{test data column}

\item{train_x}{train data column}

\item{train_y}{train data column with values you are trying to predict}

\item{k}{k nearest neighbours to consider}
}
\value{
vector with predictions
}
\description{
calculates predictions for each value in test data
}
\examples{
Using the palmerpenguins dataset

predictions <- knn_test(test$flipper_length_mm, train$flipper_length_mm, train$species, 3)

}
