% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logreg_confusion_matrix.R
\name{logreg_confusion_matrix}
\alias{logreg_confusion_matrix}
\title{logreg_confusion_matrix}
\usage{
logreg_confusion_matrix(
  log_reg,
  test,
  var,
  prob = 0.5,
  type = "response",
  margin = 1
)
}
\arguments{
\item{log_reg}{the regression}

\item{test}{the test dataframe}

\item{var}{the variable you are trying to predict}

\item{prob}{(default: 0.5) probabilites ABOVE this are predicted TRUE (x > prob => TRUE)}

\item{type}{(default: TRUE) parameter in the predict() function from the stats package}

\item{margin}{(default : 1) parameter in the prop.table() function from base R}
}
\value{
returns a table in the form of a confusion matrix
}
\description{
Returns a confusion matrix from a logistic regression
}
\examples{
Predicting default in the Default dataset from the ISLR2 package

regression <- glm(default ~ income + balance, data = train, family = 'binomial')
logreg_confusion_matrix(regression, test, 'default')

}
