% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{model_bot}
\alias{model_bot}
\title{Runs modelbot}
\usage{
model_bot(new_session = FALSE)
}
\arguments{
\item{new_session}{Logical. If \code{TRUE}, starts a new chat session. If \code{FALSE}
(the default), restores the previous chat session (only within the current
R process).}
}
\description{
Runs modelbot
}
