% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databot-package.R
\docType{package}
\name{modelbot-package}
\alias{modelbot}
\alias{modelbot-package}
\title{modelbot: Frontend For Predictive Modeling With 'tidymodels'}
\description{
A Shiny application that provides an AI-powered interface for predictive modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/modelbot}
  \item Report bugs at \url{https://github.com/simonpcouch/modelbot/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co}

Authors:
\itemize{
  \item Joe Cheng \email{joe@posit.co}
}

}
\keyword{internal}
