% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive-package.R
\docType{package}
\name{predictive-package}
\alias{predictive-package}
\title{predictive: Frontend For Predictive Modeling With 'tidymodels'}
\description{
A Shiny application that provides an AI-powered interface for predictive modeling.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonpcouch/predictive}
  \item Report bugs at \url{https://github.com/simonpcouch/predictive/issues}
}

}
\author{
\strong{Maintainer}: Simon Couch \email{simon.couch@posit.co}

Authors:
\itemize{
  \item Joe Cheng \email{joe@posit.co}
}

}
\keyword{internal}
