% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{predictive}
\alias{predictive}
\title{Frontend for tidy modeling}
\usage{
predictive(new_session = FALSE)
}
\arguments{
\item{new_session}{Logical. If \code{TRUE}, starts a new chat session. If \code{FALSE}
(the default), restores the previous chat session (only within the current
R process).}
}
\description{
predictive is an agentic frontend for predictive modeling with tidymodels.
The chat application will guide you through the exploratory data analysis
and model development process, running R code and launching modeling
experiments on your behalf.
}
\seealso{
The introductory vignette \code{vignette("predictive", package = "predictive")}
contains more information on the workflow and design of this application.
}
