% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{bootstrap function}
\usage{
bootstrap(B, x, replacement = TRUE)
}
\arguments{
\item{B}{number of samples}

\item{x}{sampling variable}

\item{replacement}{(by default TRUE) whether or not to use replacement}
}
\value{
sampling distribution
}
\description{
bootstrap function
}
\examples{
estimating population distribution of mpg using the Auto dataset from the ISLR2 package

sampling_distribution <- bootstrap(1000, Auto$mpg)

}
