% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_train_test.R
\name{split_train_test}
\alias{split_train_test}
\title{split_train_test}
\arguments{
\item{df}{dataframe you want to split}

\item{train_frac}{(default: 0.7) the fraction of the data you want in training data}
}
\value{
list containing two dataframes: train and test
}
\description{
Split dataframe into training and test data
}
\examples{
Splitting the dataframe auto, with 80\% in train and 20\% in test

result <- split_train_test(auto, train_frac = 0.8)
train <- result$train
test <- result$test

}
